+++ src/include/basic_io.h
@@ -25,8 +25,8 @@ public:
         refresh();
     }
 
-    uint64_t modify(const char * buffer, uint64_t length, uint64_t start);
-    uint64_t copy(char * buffer, uint64_t length, uint64_t start);
+    uint32_t modify(const char * buffer, uint32_t length, uint32_t start);
+    uint32_t copy(char * buffer, uint32_t length, uint32_t start);
 
     block_t & operator = (const block_t &) = delete;
 };
